+++ romsearch/modules/__init__.py
@@ -1,6 +1,7 @@
 from .datparser import DATParser
 from .dupeparser import DupeParser
 from .gamefinder import GameFinder
+from .rahasher import RAHasher
 from .romchooser import ROMChooser
 from .romdownloader import ROMDownloader
 from .rommover import ROMMover
@@ -11,6 +12,7 @@ __all__ = [
     "DATParser",
     "DupeParser",
     "GameFinder",
+    "RAHasher",
     "ROMChooser",
     "ROMDownloader",
     "ROMMover",